% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tcxr.R
\name{parse_lap}
\alias{parse_lap}
\title{Parse a Lap from a TCX File}
\usage{
parse_lap(lap, activity_id = NA_integer_, lap_id = NA_integer_)
}
\arguments{
\item{lap}{An XML node representing a lap in a TCX file.}

\item{activity_id}{Integer activity index.}

\item{lap_id}{Integer lap index within activity.}
}
\value{
A list with:
  - summary: dataframe containing lap summary
  - raw: dataframe of lap trackpoints
}
\description{
Extracts summary data from a lap and also returns raw trackpoint data.
}
