% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tcxr.R
\name{parse_trackpoint}
\alias{parse_trackpoint}
\title{Parse a Trackpoint from a TCX File}
\usage{
parse_trackpoint(
  trackpoint,
  activity_id = NA_integer_,
  lap_id = NA_integer_,
  trackpoint_id = NA_integer_
)
}
\arguments{
\item{trackpoint}{An XML node representing a trackpoint.}

\item{activity_id}{Integer activity index.}

\item{lap_id}{Integer lap index.}

\item{trackpoint_id}{Integer trackpoint index.}
}
\value{
A one-row dataframe of parsed trackpoint metrics.
}
\description{
Extracts raw trackpoint fields into one dataframe row.
}
