% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tcxr.R
\name{parse_lap}
\alias{parse_lap}
\title{Parse a Lap from a TCX File}
\usage{
parse_lap(lap)
}
\arguments{
\item{lap}{An XML node representing a lap in a TCX file.}
}
\value{
A dataframe containing the lap metrics.
}
\description{
Extracts data from a lap within a TCX file, including time, distance, altitude, speed, power, cadence.
}
