% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tcxr.R
\name{parse_lap}
\alias{parse_lap}
\title{Parse a lap from a TCX file}
\usage{
parse_lap(lap)
}
\arguments{
\item{lap}{An XML node representing a lap.}
}
\value{
A list containing the lap time, distance, calories burned, maximum altitude, and power values.
}
\description{
This function extracts lap details including time, distance, calories, maximum altitude, and watts.
}
