% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tcxr.R
\name{parse_trackpoint}
\alias{parse_trackpoint}
\title{Parse a trackpoint from a TCX file}
\usage{
parse_trackpoint(trackpoint)
}
\arguments{
\item{trackpoint}{An XML node representing a trackpoint.}
}
\value{
A list containing the altitude, distance, heart rate, and power output at the trackpoint.
}
\description{
This function extracts trackpoint details such as altitude, distance, heart rate, and power values.
}
