% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tcxr.R
\name{TCXRead}
\alias{TCXRead}
\title{Read and Parse a TCX File}
\usage{
TCXRead(file_path)
}
\arguments{
\item{file_path}{A character string specifying the path to the TCX file.}
}
\value{
A list containing the computed activity metrics, including the activity type.
}
\description{
Parses a TCX file to extract key activity metrics such as speed, distance, time, altitude, power, cadence, heart rate, and activity type.
}
